hello/hello.go
package hello


const spanish = "Spanish"
const englishHelloPrefix = "Hello, "
const spanishHelloPrefix = "Hola, "
const frenchHelloPrefix = "Bonjour, "
const brazilianHelloPrefix = "Ola, "

// Hello method retur string
func Hello(name string, language string) string {
	if name == "" {
		name = "World"
	}
	return greetingPrefix(language) + name
}

func greetingPrefix(language string) (prefix string) {
	switch language {
	case "french":
		prefix = frenchHelloPrefix
	case "spanish":
		prefix = spanishHelloPrefix
	case "brazil":
		prefix = brazilianHelloPrefix
	default:
		prefix = englishHelloPrefix
	}
	return
}

